Use proper numbering of the Interior elements
ParMetis only knows about interior elements, and numbers them accordingly.
However, the element list of a Dune grid will also contain Ghost elements,
and they may appear anywhere in the list of elements.  This patch adds
proper renumbering of the ParMetis output to the numbering produced by
a MultipleCodimMultipleGeomTypeMapper, which is the one expected by
the UGGrid loadBalance method.

Signed-off-by: Convex Function <329364@gehry.mathepool.rwth-aachen.de>